Grab the screen at the given QCursor::pos coordinates.
This grabs the screen at the coordinates of the cursor, or at least it
is supposed to do that.  No width or height is specified in this case.

This appears to ignore the X position if the cursor is on the
non-primary screen in a 2-screen scenario, where the non-primary screen
is to the left of the primary screen. Tested on Windows 10, Qt 5.15.2.
In this case, the screen is grabbed from the leftmost part of the
non-primary screen, and encompases the rectangle that includes all of
both screens in the x direction, but only the lower half of both screens
in the Y (starting at the Y position of the cursor - i.e. Y seems to be
interpreted properly, but X is not).